Small changes for RISC OS 64 build.
Building on RISC OS 64 isn't too hard, but I needed to change a few
functions around to be more like my library. We don't have the setlocale
yet, and the time is read through the OS_ReadMonotonicTime call instead
of the POSIX functions. Otherwise the rest of the change is largely
about making it work in the build environment.